+++ tests/controllers/controller_with_media_types_routing.py
@@ -2,7 +2,6 @@ import winter
 from winter.web import MediaType
 
 
-@winter.controller
 @winter.route('controller_with_media_types_routing/')
 class ControllerWithMediaTypesRouting:
     @winter.route_get('xml/', produces=(MediaType.APPLICATION_XML, ))