+++ tests/controllers/controller_with_path_parameters.py
@@ -18,7 +18,6 @@ class OneTwoEnumWithInt(enum.Enum):
         return cls(int(value))
 
 
-@winter.controller
 @winter.route('controller_with_path_parameters/{param1}/')
 class ControllerWithPathParameters:
     @winter.route_get('{param2}/{param3}/{param4}/{param5}/{?param6}')