+++ tests/controllers/controller_with_response_headers.py
@@ -7,7 +7,6 @@ import winter
 from winter.web import ResponseHeader
 
 
-@winter.controller
 @winter.route('with-response-headers/')
 class ControllerWithResponseHeaders: