+++ tests/controllers/no_authentication_controller.py
@@ -1,7 +1,6 @@
 import winter.web
 
 
-@winter.controller
 @winter.route('winter-no-auth')
 @winter.web.no_authentication
 class NoAuthenticationController: