+++ tests/drf/controller_with_output_template.py
@@ -5,7 +5,6 @@ import winter
 import winter_django
 
 
-@winter.controller
 @winter.route('with-output-template/')
 class ControllerWithOutputTemplate: