+++ tests/drf/controller_with_serializer.py
@@ -24,7 +24,6 @@ class SerializerWithContext(serializers.Serializer):
         return self.context['additional_data']
 
 
-@winter.controller
 @winter.route('with-serializer')
 class ControllerWithSerializer: