+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.6.5]
+
+- Deprecate @winter.web.controller
+- Delete get_component function in favor of direct Component.get_by_cls
+
 ## [9.6.4] - 2022-08-01
 ### Added 
 - winter_openapi: add support of StrEnum: now StrEnum will be recognized as Enum, not as simple string